requirements.txt
streamlit==1.43.0
langchain==0.3.20
langchain-openai==0.3.7
python-dotenv==1.0.1
